% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{covEllipses}
\alias{covEllipses}
\title{Plot ellipses using covariance matrices without access to raw data}
\usage{
covEllipses(means, covariances, x = 1, y = 2)
}
\arguments{
\item{means}{a matrix-like R object, with at least two dimensions}

\item{covariances}{a matrix-like R object, with at least two dimensions}

\item{x}{index of data for the x axis}

\item{y}{index of data for the y axis}
}
\value{
ggplot2 plot
}
\description{
Plot ellipses using covariance matrices without access to raw data
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
idData = xml2gmLdaData(xmlPath)
idMeans = idData$means \%*\% t(idData$coefficients)
covEllipses(idMeans, idData$covariances)
}
