% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2id}
\alias{gmLdaData2id}
\title{Classify unknown samples using identification data}
\usage{
gmLdaData2id(idData, data, average = TRUE)
}
\arguments{
\item{idData}{list of identification data}

\item{data}{unknown data to classify}

\item{average}{average the data or classify each row}
}
\value{
list of id data
}
\description{
Classify unknown samples using identification data
}
\examples{
idData <- xml2gmLdaData("Apis-mellifera-lineages.dw.xml")
id <- gmLdaData2id(idData, unknownDat)
}
