% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2xml}
\alias{gmLdaData2xml}
\title{Create XML document with classification data}
\usage{
gmLdaData2xml(inData, grVec, file, prototype = "")
}
\arguments{
\item{inData}{a matrix-like R object, with at least six columns}

\item{grVec}{a vector with grouping variable}

\item{file}{a character string with file name}

\item{prototype}{an optional character string with prototype file name. It is used by IdentiFly to describe landmarks.}
}
\description{
Create XML document with classification data
}
\examples{
data(lineages)
grVec = lineages$lineage
coordinates = lineages[,-1] # remove the first column
gmLdaData2xml(coordinates, grVec, "apis-mellifera-lineage.dw.xml")
idData = xml2gmLdaData("apis-mellifera-lineage.dw.xml")
id = gmLdaData2id(idData, coordinates, average = FALSE)
id$plot
}
