% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineages}
\alias{lineages}
\title{Honey bee lineages}
\format{
A data frame with 187 rows and 39 columns:
\describe{
  \item{lineage}{Lineage name}
  \item{x1}{x Cartesian coordinate of landmark 1}
  \item{y1}{y Cartesian coordinate of landmark 1}
  \item{x2}{x Cartesian coordinate of landmark 2}
  \item{y2}{y Cartesian coordinate of landmark 2}
}
}
\source{
https://zenodo.org/records/7567336
}
\usage{
data(lineages)
}
\description{
Coordinates of 19 landmarks on honey bee (*Apis mellifera*) wings.
The coordinates were scaled and aligned. Each row represents one colony.
There are 187 colonies representing 4 lineages: A, C, M and O.
}
\keyword{datasets}
