% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{reorder2D}
\alias{reorder2D}
\title{Reorder 2D landmarks using column names}
\usage{
reorder2D(inData, newOrder)
}
\arguments{
\item{inData}{a matrix-like R object, with two dimensions}

\item{newOrder}{a numerical vector with new order of landmarks}
}
\value{
a matrix with reordered columns
}
\description{
Reorder 2D landmarks using column names
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
idData = xml2gmLdaData(xmlPath)
p <- 19 # number of landmarks
k <- 2 # number of dimensions
reference = idData$reference
reference3D <- geomorph::arrayspecs(reference, p, k)
geomorph::plotAllSpecimens(reference3D, label=TRUE)
Bustamente2IdentiFly =
c(18, 17, 15, 14, 16, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 19)
reference = reorder2D(reference, Bustamente2IdentiFly)
reference3D <- geomorph::arrayspecs(reference, p, k)
geomorph::plotAllSpecimens(reference3D, label=TRUE)
}
