% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{xml2gmLdaData}
\alias{xml2gmLdaData}
\title{Read classification data from XML file}
\usage{
xml2gmLdaData(path)
}
\arguments{
\item{path}{file name with xml document}
}
\value{
list of classification data
}
\description{
Read classification data from XML file
}
\examples{
idData = xml2gmLdaData("https://zenodo.org/record/10512712/files/apis-mellifera-queens-workers-drones.dw.xml")
unknown <- read.csv("https://zenodo.org/record/8071014/files/IN-raw-coordinates.csv")
unknown <- data.frame(unknown, row.names = 1)  # move column 1 to row names
id = gmLdaData2id(idData, unknown, average = TRUE)
id$plot
id$id
}
