% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{xml2id}
\alias{xml2id}
\title{Classify unknown data using data from XML file}
\usage{
xml2id(path, data, average = TRUE)
}
\arguments{
\item{path}{a file name with xml document}

\item{data}{a matrix of unknown data to classify}

\item{average}{an optionally Boolean variable indicating if rows are averaged}
}
\value{
list of classification data
\itemize{
  \item reference - matrix with reference configuration.
  \item means - matrix with class means.
  \item covariances - list of covariance matrices.
  \item coefficients - matrix with LDA coefficients.
}
}
\description{
Classify unknown data using data from XML file
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
data(lineages)
unknownData <- lineages[,-1]
id <- xml2id(xmlPath, unknownData)
}
