% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{covEllipses}
\alias{covEllipses}
\title{Plot ellipses for the first two LD using data from covariance matrices}
\usage{
covEllipses(means, covariances, x = 1, y = 2)
}
\arguments{
\item{means}{a matrix-like R object, with at least two dimensions}

\item{covariances}{a matrix-like R object, with at least two dimensions}

\item{x}{index of data for the x axis}

\item{y}{index of data for the y axis}
}
\value{
ggplot2 plot
}
\description{
Plot ellipses for the first two LD using data from covariance matrices
}
\examples{
path = "https://zenodo.org/record/10512712/files/apis-mellifera-queens-workers-drones.dw.xml"
idData = xml2gmLdaData(path)
idMeans = idData$means \%*\% t(idData$coefficients)
covEllipses(idMeans, idData$covariances)
}
