% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2id}
\alias{gmLdaData2id}
\title{Classify unknown samples using identification data}
\usage{
gmLdaData2id(idData, data, average = TRUE)
}
\arguments{
\item{idData}{a list of identification data}

\item{data}{a matrix of unknown data to classify}

\item{average}{an optionally Boolean variable indicating if rows are averaged}
}
\value{
list of id data
\itemize{
  \item id - data frame with 2 columns: group and probability.
  \item P - data frame with classification probabilities.
  \item MD2 - data frame with squared Mahalanobis distances.
  \item LD - data frame with linear discriminant scores.
  \item plot - ggplot2 plot.
}
}
\description{
Classify unknown samples using identification data
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
idData = xml2gmLdaData(xmlPath)
data(lineages)
unknownData <- lineages[,-1]
id <- gmLdaData2id(idData, unknownData)
}
