% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2id}
\alias{gmLdaData2id}
\title{Classify unknown samples using identification data}
\usage{
gmLdaData2id(idData, data, average = TRUE)
}
\arguments{
\item{idData}{List of identification data}

\item{data}{Matrix of unknown data to classify}

\item{average}{Optionally classify all rows or averyge only}
}
\value{
list of id data
}
\description{
Classify unknown samples using identification data
}
\examples{
idData <- xml2gmLdaData("https://zenodo.org/record/10512712/files/apis-mellifera-queens-workers-drones.dw.xml")
unknownDat <- read.csv("https://zenodo.org/record/8071014/files/IN-raw-coordinates.csv")
unknownDat <- data.frame(unknownDat, row.names = 1)  # move column 1 to row names
id <- gmLdaData2id(idData, unknownDat)
}
