% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2id}
\alias{gmLdaData2id}
\title{Classify unknown samples using identification data}
\usage{
gmLdaData2id(idData, data, average = TRUE)
}
\arguments{
\item{idData}{List of identification data}

\item{data}{Matrix of unknown data to classify}

\item{average}{Optionally classify all rows or averyge only}
}
\value{
list of id data
}
\description{
Classify unknown samples using identification data
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
idData = xml2gmLdaData(xmlPath)
data(lineages)
unknownData <- lineages[,-1]
id <- gmLdaData2id(idData, unknownData)
}
