% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{gmLdaData2xml}
\alias{gmLdaData2xml}
\title{Create DW.XML file with classification data}
\usage{
gmLdaData2xml(inData, grouping)
}
\arguments{
\item{grouping}{a vector with grouping variable}

\item{data}{a matrix-like R object, with at least two dimensions}
}
\value{
XML document
}
\description{
Create DW.XML file with classification data
}
\examples{
outXML <- gmLdaData2xml(inData, grVec)
}
