% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{xml2gmLdaData}
\alias{xml2gmLdaData}
\title{Read classification data from XML file}
\usage{
xml2gmLdaData(path)
}
\arguments{
\item{path}{file name with xml document}
}
\value{
list of classification data
\itemize{
  \item reference - matrix with reference configuration.
  \item means - matrix with class means.
  \item covariances - list of covariance matrices.
  \item coefficients - matrix with LDA coefficients.
}
}
\description{
Read classification data from XML file
}
\examples{
xmlPath = system.file("extdata",
                      "apis-mellifera-queens-workers-drones.dw.xml",
                      package="IdentiFlyR")
idData = xml2gmLdaData(xmlPath)
data(lineages)
unknownData <- lineages[,-1]
id = gmLdaData2id(idData, unknownData, average = FALSE)
id$plot
}
