% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{xml2gmLdaData}
\alias{xml2gmLdaData}
\title{Read classification data from XML file}
\usage{
xml2gmLdaData(path)
}
\arguments{
\item{path}{file name with xml document}
}
\value{
list of classification data
}
\description{
Read classification data from XML file
}
\examples{
idDataList <- xml2gmLdaData("apis-mellifera-lineage.dw.xml")
}
