% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentiFlyR.R
\name{xml2id}
\alias{xml2id}
\title{Classify unknown samples using XML data}
\usage{
xml2id(path, data, average = TRUE)
}
\arguments{
\item{path}{file name with xml document}

\item{data}{unknown data to classify}

\item{average}{average the data or classify each row}
}
\value{
list of id data
}
\description{
Classify unknown samples using XML data
}
\examples{
idData <- xml2id("Apis-mellifera-lineages.dw.xml", unknownDat)
}
